code/requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
jellyfish==0.6.1
kiwisolver==1.1.0
matplotlib==2.2.4
nltk==3.4.4
numpy==1.16.4
pyenchant==2.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytypo==0.3.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
subprocess32==3.5.4
textblob==0.15.3
